Update x264 to the latest stable commit
The git url is changed to the upstream repository, because the "mirror"
repository used until now does not seem to be an official mirror and has not
been updated for more than a year.

The flatpak external data checker extracted the latest stable commit id from
the upstream repository, but the outdated "mirror" repository was used to fetch
the source, so that builds with an updated x264 failed because the requested
commit did not exist in the source repository.